Rename StudentRepository to StudentRepo and update usage
Renamed StudentRepository.java to StudentRepo.java and moved it to the model package. Updated all references in AddStudentController and HomeController to use StudentRepo. Also added search functionality in HomeController to filter students by ID.
