+++ src/main/java/com/school/studentmanagementfx/controller/HomeController.java
@@ -2,7 +2,9 @@ package com.school.studentmanagementfx.controller;
 
 import com.school.studentmanagementfx.helper.CreateWindow;
 import com.school.studentmanagementfx.model.Student;
-import com.school.studentmanagementfx.repository.StudentRepository;
+import com.school.studentmanagementfx.model.StudentRepo;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.geometry.Insets;
@@ -56,12 +58,19 @@ public class HomeController {
         yearTableColumn.setCellValueFactory(cellData -> cellData.getValue().getYear());
         addDetailButton();
 
-        studentsTableView.setItems(StudentRepository.getStudents());
+        studentsTableView.setItems(StudentRepo.getStudents());
     }
 
     @FXML
     private void onSearchStudentAction() {
-
+        String queryId = searchTextField.getText().trim().toLowerCase();
+        ObservableList<Student> filtered = FXCollections.observableArrayList();
+        for (Student s : StudentRepo.getStudents()) {
+            if (s.getId().get().toLowerCase().contains(queryId)) {
+                filtered.add(s);
+                System.out.println("Student found: " + s.getName());
+            }
+        }
     }
 
     @FXML
@@ -105,10 +114,10 @@ public class HomeController {
         Student student3 = new Student("1", "fucking shit", "19", "12-31-2005", "Panacan Davao City", "BSIT", "2nd year", "johnreyrabosa.f@gmail.com");
         Student student4 = new Student("1", "John Rey Rabosa", "19", "12-31-2005", "Panacan Davao City", "BSIT", "2nd year", "johnreyrabosa.f@gmail.com");
         Student student5 = new Student("1", "John Rey Rabosa", "19", "12-31-2005", "Panacan Davao City", "BSIT", "2nd year", "johnreyrabosa.f@gmail.com");
-        StudentRepository.getStudents().add(student1);
-        StudentRepository.getStudents().add(student2);
-        StudentRepository.getStudents().add(student3);
-        StudentRepository.getStudents().add(student4);
-        StudentRepository.getStudents().add(student5);
+        StudentRepo.getStudents().add(student1);
+        StudentRepo.getStudents().add(student2);
+        StudentRepo.getStudents().add(student3);
+        StudentRepo.getStudents().add(student4);
+        StudentRepo.getStudents().add(student5);
     }
 }
